requirements.txt
django
django-bootstrap-v5
django[bcrypt]
django-crispy-forms
crispy-bootstrap5
requests
